requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
chardet==4.0.0
cycler==0.10.0
feedparser==6.0.2
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
